plus

+++ mandatory/my_utils.c
@@ -15,9 +15,9 @@
 void	my_mlx_pixel_put(t_img_data *img, int x, int y, int color)
 {
 	char			*dst;
-	unsigned int	img_width;
-	unsigned int	pixel_index;
-	unsigned int	pixel_size;
+	unsigned long	img_width;
+	unsigned long	pixel_index;
+	unsigned long	pixel_size;
 
 	img_width = img->line_length;
 	pixel_size = img->bits_per_pixel / 8;
